Update readme and add azUtils
