+++ src/utils/dotnetUtils.ts
@@ -51,27 +51,4 @@ export namespace dotnetUtils {
             throw new Error(message);
         }
     }
-
-    export async function isAzInstalled(): Promise<boolean> {
-        try {
-            await cpUtils.executeCommand(undefined, undefined, 'az', '--version');
-            return true;
-        } catch (error) {
-            return false;
-        }
-    }
-
-    export async function checkAzInstalled(): Promise<void> {
-        if (!await isAzInstalled()) {
-            const message: string = localize('azNotInstalled', 'You must have the AZ CLI installed to perform this operation.');
-
-            vscode.window.showErrorMessage(message, DialogResponses.learnMore).then(async (result) => {
-                if (result === DialogResponses.learnMore) {
-                    await openUrl('https://docs.microsoft.com/en-us/cli/azure/install-azure-cli?view=azure-cli-latest');
-                }
-            });
-
-            throw new Error(message);
-        }
-    }
 }
